+++ helpers/file_helper.rb
@@ -2,24 +2,26 @@
 
 require 'open-uri'
 require 'tempfile'
+require 'net/http'
 
 # class with methods for working with files
 class FileHelper
-  # delete all files and folders from dir
-  def self.clear_dir(dir)
-    FileUtils.rm_rf("#{dir}/.", secure: true)
-  end
-
   # create test file in folder
   def self.create_file(filepath)
     FileUtils.touch(filepath)
   end
 
+  def self.create_tmp_dir
+    tmp_dir = File.join(StaticData::TMP_FOLDER, "tmp_#{Time.now.to_i}_#{rand(10_000)}")
+    FileUtils.mkdir_p(tmp_dir)
+    tmp_dir
+  end
+
   # @param file_name [String] Accepts file name
   # @return [String] will return the generated name
-  def self.file_rename(file_name)
-    file_new_name = Time.now.nsec.to_s + File.extname(file_name)
-    File.rename("files_tmp/#{file_name}", "files_tmp/#{file_new_name}")
+  def self.file_rename(file_path)
+    file_new_name = Time.now.nsec.to_s + File.extname(File.basename(file_path))
+    File.rename(file_path, File.join(File.dirname(file_path), file_new_name))
     file_new_name
   end
 
@@ -36,14 +38,7 @@ class FileHelper
     end
   end
 
-  # Download temp file and return it location
-  # @param file_url [String] url
-  # @return [String] path to file
-  def self.download_file(url)
-    data = URI.parse(url).open.read
-    file = Tempfile.new('convert-service-file')
-    file.write(data.force_encoding('UTF-8'))
-    file.close
-    file.path
+  def self.download_file(url, file_path)
+    File.binwrite(file_path, Net::HTTP.get_response(URI(url)).body)
   end
 end